pyproject.toml
[project]
name = "qb-proxy-service"
version = "0.1.0"
description = "FastAPI microservice proxying QuickBooks API"
authors = [
    {name = "Your Name",email = "you@domain.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "authlib (>=1.6.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
